fix: 🐛 change upload and addSubscription code with normal
